feat: add serial and sequence management hooks
