+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import React, { useState, useEffect, useCallback, useRef } from 'react';
+import React, { useEffect, useState } from 'react';
 import Header from '@/components/dashboard/header';
 import SensorPanel from '@/components/dashboard/sensor-panel';
 import ValveDisplay from '@/components/dashboard/valve-display';
@@ -8,81 +8,37 @@ import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import SequencePanel from '@/components/dashboard/sequence-panel';
 import DataChartPanel from '@/components/dashboard/data-chart-panel';
 import TerminalPanel from '@/components/dashboard/terminal-panel';
-import { useToast } from "@/hooks/use-toast";
-import { AppConfig } from '@/types';
-
-// Data types
-export interface SensorData {
-  pt1: number;
-  pt2: number;
-  pt3: number;
-  pt4: number;
-  flow1: number;
-  flow2: number;
-  tc1: number;
-  timestamp: number;
-}
-
-export type ValveState = 'OPEN' | 'CLOSED' | 'OPENING' | 'CLOSING' | 'ERROR';
-
-export interface Valve {
-  id: number;
-  name: string;
-  state: ValveState;
-  lsOpen: boolean;
-  lsClosed: boolean;
-}
-
-const initialValves: Valve[] = [
-  { id: 1, name: 'Ethanol Main', state: 'CLOSED', lsOpen: false, lsClosed: false },
-  { id: 2, name: 'N2O Main', state: 'CLOSED', lsOpen: false, lsClosed: false },
-  { id: 3, name: 'Ethanol Purge', state: 'CLOSED', lsOpen: false, lsClosed: false },
-  { id: 4, name: 'N2O Purge', state: 'CLOSED', lsOpen: false, lsClosed: false },
-  { id: 5, name: 'Pressurant Fill', state: 'CLOSED', lsOpen: false, lsClosed: false },
-  { id: 6, name: 'System Vent', state: 'CLOSED', lsOpen: false, lsClosed: false },
-  { id: 7, name: 'Igniter Fuel', state: 'CLOSED', lsOpen: false, lsClosed: false },
-];
-
-
-const initialSensorData: SensorData = {
-  pt1: 0,
-  pt2: 0,
-  pt3: 0,
-  pt4: 0,
-  flow1: 0,
-  flow2: 0,
-  tc1: 0,
-  timestamp: 0,
-};
-
-function isSensorDataKey(key: string): key is keyof SensorData {
-  return key in initialSensorData;
-}
-
-const MAX_CHART_DATA_POINTS = 100;
-const PRESSURE_LIMIT = 850; // PSI
+import { useToast } from '@/hooks/use-toast';
+import { useSerialManager } from '@/hooks/useSerialManager';
+import { useSequenceManager } from '@/hooks/useSequenceManager';
 
 export default function Home() {
   const { toast } = useToast();
-  const [sensorData, setSensorData] = useState<SensorData | null>(null);
-  const sensorDataRef = useRef<SensorData | null>(sensorData);
-  const [chartData, setChartData] = useState<SensorData[]>([]);
-  const [valves, setValves] = useState<Valve[]>(initialValves);
-  const [connectionStatus, setConnectionStatus] = useState<'connected' | 'disconnected' | 'connecting'>('disconnected');
-  const [sequenceLogs, setSequenceLogs] = useState<string[]>(['System standby. Select a sequence to begin.']);
-  const [activeSequence, setActiveSequence] = useState<string | null>(null);
-  
-  const [serialPorts, setSerialPorts] = useState<string[]>([]);
-  const [selectedPort, setSelectedPort] = useState<string>('');
-  const [isLogging, setIsLogging] = useState(false);
-  const [appConfig, setAppConfig] = useState<AppConfig | null>(null);
+  const {
+    sensorData,
+    chartData,
+    valves,
+    connectionStatus,
+    serialPorts,
+    selectedPort,
+    setSelectedPort,
+    handleConnect,
+    handleValveChange,
+    sendCommand,
+    setLogger,
+    setSequenceHandler,
+  } = useSerialManager();
 
-  const sequenceTimeoutRef = useRef<NodeJS.Timeout[]>([]);
-  const emergencyShutdownTriggered = useRef(false);
-  const ignitionPhase = useRef<'idle' | 'igniter' | 'main_stage'>('idle');
+  const { sequenceLogs, activeSequence, handleSequence, addLog } = useSequenceManager(sendCommand);
+
+  const [isLogging, setIsLogging] = useState(false);
 
   useEffect(() => {
-    // Zoom control
+    setLogger(addLog);
+    setSequenceHandler(handleSequence);
+  }, [addLog, handleSequence, setLogger, setSequenceHandler]);
+
+  useEffect(() => {
     const handleWheel = (event: WheelEvent) => {
       if (event.ctrlKey) {
         event.preventDefault();
@@ -119,202 +75,10 @@ export default function Home() {
   }, []);
 
   useEffect(() => {
-    const getPorts = async () => {
-      try {
-        const ports = await window.electronAPI.getSerialPorts();
-        setSerialPorts(ports);
-        if (ports.length > 0) {
-          setSelectedPort(ports[0]);
-        }
-      } catch (error) {
-        console.error(error);
-        toast({ title: "Connection Error", description: "Failed to list serial ports.", variant: "destructive" });
-      }
-    };
-    getPorts();
-
-    const loadConfig = async () => {
-      try {
-        const cfg = await window.electronAPI.getConfig();
-        setAppConfig(cfg);
-      } catch (error) {
-        console.error(error);
-        toast({ title: "Configuration Error", description: "Failed to load configuration.", variant: "destructive" });
-      }
-    };
-    loadConfig();
-
-    const handleSerialData = (data: string) => {
-      addLog(`Received: ${data}`);
-      // Assuming data format is "key1:value1,key2:value2,..."
-      const parts = data.split(',');
-      const newData: Partial<SensorData> = {};
-      const newValveStates: Partial<Record<number, Partial<Valve>>> = {};
-
-      parts.forEach(part => {
-        const [key, rawValue] = part.split(':');
-        if (!key || !rawValue) return;
-        const value = rawValue.trim(); // Add .trim() here
-
-        // Update sensor data
-        if (isSensorDataKey(key)) {
-            newData[key] = parseFloat(value);
-        }
-
-        // Update valve states based on limit switches
-        const match = key.match(/V(\d)(LS_OPEN|LS_CLOSED)/);
-        if (match) {
-            const valveId = parseInt(match[1]);
-            const lsType = match[2];
-            const lsValue = value === '1';
-            if (!newValveStates[valveId]) newValveStates[valveId] = {};
-            if (lsType === 'LS_OPEN') newValveStates[valveId]!.lsOpen = lsValue;
-            if (lsType === 'LS_CLOSED') newValveStates[valveId]!.lsClosed = lsValue;
-        }
-
-        // REMOVED: Handle motor state updates from Arduino (e.g., "D1C0:OPEN")
-        // const motorStateMatch = key.match(/D(\d)C(\d)/);
-        // if (motorStateMatch) {
-        //     const driver = parseInt(motorStateMatch[1]);
-        //     const channel = parseInt(motorStateMatch[2]);
-        //     const receivedState = value as ValveState; // "OPEN" or "CLOSED"
-        //     console.log(`Received motor state: ${receivedState}`); // 추가된 로그
-
-        //     // Find the valve that maps to this driver and channel
-        //     const valveIdToUpdate = Object.keys(valveMappings).find(valveId => {
-        //         const mapping = valveMappings[parseInt(valveId)];
-        //         return mapping.servoIndex === servoIndex;
-        //     });
-
-        //     if (valveIdToUpdate) {
-        //         const id = parseInt(valveIdToUpdate);
-        //         if (!newValveStates[id]) newValveStates[id] = {};
-        //         newValveStates[id]!.state = receivedState;
-        //     }
-        // }
-      });
-      
-      if (Object.keys(newData).length > 0) {
-        const updatedSensorData = { ...sensorDataRef.current, ...newData, timestamp: Date.now() } as SensorData;
-        setSensorData(updatedSensorData);
-
-        // Only add to chart if tc1 is a valid number
-        if (typeof updatedSensorData.tc1 === 'number') {
-            setChartData(prev => [...prev.slice(-MAX_CHART_DATA_POINTS + 1), updatedSensorData]);
-        }
-        
-        if ((updatedSensorData.pt1 > PRESSURE_LIMIT || updatedSensorData.pt2 > PRESSURE_LIMIT) && !emergencyShutdownTriggered.current) {
-          addLog(`!!! CRITICAL PRESSURE DETECTED (PT1: ${updatedSensorData.pt1.toFixed(0)} PSI, PT2: ${updatedSensorData.pt2.toFixed(0)} PSI) !!!`);
-          handleSequence("Emergency Shutdown");
-          emergencyShutdownTriggered.current = true;
-        } else if (updatedSensorData.pt1 < PRESSURE_LIMIT && updatedSensorData.pt2 < PRESSURE_LIMIT) {
-            emergencyShutdownTriggered.current = false;
-        }
-      }
-
-      if(Object.keys(newValveStates).length > 0) {
-        setValves(prevValves => prevValves.map(v => {
-            const updates = newValveStates[v.id];
-            if (!updates) return v;
-
-            const newState = { ...v, ...updates };
-
-            // If a limit switch is active, update the main state accordingly
-            if (newState.lsOpen) {
-                newState.state = 'OPEN';
-            } else if (newState.lsClosed) {
-                newState.state = 'CLOSED';
-            } 
-            // Note: If no limit switch is active, we keep the optimistic state set by handleValveChange
-
-            return newState;
-        }));
-      }
-    };
-
-    const handleSerialError = (error: string) => {
-        addLog(`SERIAL ERROR: ${error}`);
-        toast({ title: "Serial Port Error", description: error, variant: "destructive" });
-        setConnectionStatus('disconnected');
-    };
-
-    const cleanupSerialData = window.electronAPI.onSerialData(handleSerialData);
-    const cleanupSerialError = window.electronAPI.onSerialError(handleSerialError);
-
-    return () => {
-      cleanupSerialData();
-      cleanupSerialError();
-      sequenceTimeoutRef.current.forEach(clearTimeout);
-    };
-  }, []);
-
-  useEffect(() => {
-    sensorDataRef.current = sensorData;
-  }, [sensorData]);
-
-  useEffect(() => {
     window.electronAPI.onLogCreationFailed(() => {
-      toast({ title: "Logging Error", description: "Failed to create log file.", variant: "destructive" });
+      toast({ title: 'Logging Error', description: 'Failed to create log file.', variant: 'destructive' });
     });
   }, [toast]);
-
-  const handleConnect = async () => {
-    if (connectionStatus === 'connected') {
-      await window.electronAPI.disconnectSerial();
-      setConnectionStatus('disconnected');
-      addLog(`Disconnected from ${selectedPort}.`);
-    } else {
-      if (!selectedPort) {
-        toast({ title: "Connection Error", description: "Please select a serial port.", variant: "destructive" });
-        return;
-      }
-      setConnectionStatus('connecting');
-      addLog(`Connecting to ${selectedPort}...`);
-      const success = await window.electronAPI.connectSerial(selectedPort);
-      if (success) {
-        setConnectionStatus('connected');
-        addLog(`Successfully connected to ${selectedPort}.`);
-      } else {
-        setConnectionStatus('disconnected');
-        addLog(`Failed to connect to ${selectedPort}.`);
-      }
-    }
-  };
-
-  const sendCommand = (cmd: string) => {
-    if (connectionStatus !== 'connected') {
-        toast({ title: "Not Connected", description: "Must be connected to a serial port to send commands.", variant: "destructive" });
-        return;
-    }
-    window.electronAPI.sendToSerial(cmd);
-    addLog(`Sent: ${cmd}`);
-  };
-
-  const handleValveChange = useCallback((valveId: number, targetState: 'OPEN' | 'CLOSED') => {
-    const valve = valves.find(v => v.id === valveId);
-    if (!valve) return;
-
-    const mapping = appConfig?.valveMappings?.[valve.name];
-    if (!mapping) {
-        toast({ title: "Command Error", description: `No servo mapping for valve ${valve.name}.`, variant: "destructive" });
-        return;
-    }
-
-    const command = `V,${mapping.servoIndex},${targetState === 'OPEN' ? 'O' : 'C'}`;
-    sendCommand(command);
-
-    // Optimistically update UI to show transition
-    setValves(prevValves => prevValves.map(v => 
-      v.id === valveId 
-        ? { ...v, state: targetState } // Directly set to targetState
-        : v
-    ));
-  }, [valves, connectionStatus, toast, appConfig]);
-  
-  const addLog = (message: string) => {
-    const timestamp = new Date().toLocaleTimeString();
-    setSequenceLogs(prev => [...prev, `[${timestamp}] ${message}`]);
-  };
 
   const handleLoggingToggle = () => {
     if (isLogging) {
@@ -323,64 +87,6 @@ export default function Home() {
     } else {
       window.electronAPI.startLogging();
       setIsLogging(true);
-    }
-  };
-
-  const clearAndRunSequence = (name: string, steps: { message: string, delay: number, action?: () => void }[]) => {
-    setActiveSequence(name);
-    setSequenceLogs([]);
-    sequenceTimeoutRef.current.forEach(clearTimeout);
-    sequenceTimeoutRef.current = [];
-    
-    let cumulativeDelay = 0;
-    
-    addLog(`Initiating sequence: ${name}`);
-
-    steps.forEach((step, index) => {
-      cumulativeDelay += step.delay;
-      const timeout = setTimeout(() => {
-        addLog(step.message);
-        step.action?.();
-        if (index === steps.length - 1) {
-            addLog(`Sequence ${name} complete.`);
-            setActiveSequence(null);
-            ignitionPhase.current = 'idle';
-        }
-      }, cumulativeDelay);
-      sequenceTimeoutRef.current.push(timeout);
-    });
-  };
-
-  const handleSequence = (sequenceName: string) => {
-    if (activeSequence) {
-        toast({
-            title: "Sequence in Progress",
-            description: `Cannot start "${sequenceName}" while "${activeSequence}" is running.`,
-            variant: "destructive",
-        });
-        return;
-    }
-
-    // All sequence actions now just send commands
-    switch (sequenceName) {
-        case "Ignition Sequence":
-            ignitionPhase.current = 'idle';
-            clearAndRunSequence("Ignition Sequence", [
-                { message: "Sending command: IGNITION_SEQUENCE_START", delay: 500, action: () => sendCommand("SEQ_IGNITION_START") },
-            ]);
-            break;
-        case "Emergency Shutdown":
-            ignitionPhase.current = 'idle';
-            clearAndRunSequence("Emergency Shutdown", [
-                { message: "Sending command: EMERGENCY_SHUTDOWN", delay: 100, action: () => sendCommand("SEQ_SHUTDOWN") },
-            ]);
-            break;
-        default:
-             clearAndRunSequence(sequenceName, [
-                { message: `Running diagnostics for ${sequenceName}...`, delay: 1000, action: () => sendCommand(`DIAG_${sequenceName.toUpperCase().replace(' ', '_')}`) },
-                { message: "Diagnostics complete.", delay: 2000 },
-             ]);
-            break;
     }
   };
 
@@ -421,7 +127,6 @@ export default function Home() {
             <SequencePanel onSequence={handleSequence} activeSequence={activeSequence} />
             <TerminalPanel logs={sequenceLogs} activeSequence={activeSequence} />
           </div>
-
         </div>
       </main>
     </div>
