+++ src/types/index.ts
@@ -8,3 +8,30 @@ export interface AppConfig {
   };
   valveMappings: Record<string, ValveMappingEntry>;
 }
+
+// Sensor and valve related types
+export interface SensorData {
+  pt1: number;
+  pt2: number;
+  pt3: number;
+  pt4: number;
+  flow1: number;
+  flow2: number;
+  tc1: number;
+  timestamp: number;
+}
+
+export type ValveState =
+  | 'OPEN'
+  | 'CLOSED'
+  | 'OPENING'
+  | 'CLOSING'
+  | 'ERROR';
+
+export interface Valve {
+  id: number;
+  name: string;
+  state: ValveState;
+  lsOpen: boolean;
+  lsClosed: boolean;
+}
